lib/requirements.txt
pandoc
python-frontmatter
pyyaml
requests
unidecode
